[RW-489] Rename reliefweb_docstore to reliefweb_files
