+++ config/core.extension.yml
@@ -60,10 +60,10 @@ module:
   reliefweb_bookmarks: 0
   reliefweb_contact: 0
   reliefweb_disaster_map: 0
-  reliefweb_docstore: 0
   reliefweb_dsr: 0
   reliefweb_entities: 0
   reliefweb_fields: 0
+  reliefweb_files: 0
   reliefweb_form: 0
   reliefweb_guidelines: 0
   reliefweb_homepage: 0
