+++ config/field.field.node.report.field_file.yml
@@ -6,7 +6,7 @@ dependencies:
     - field.storage.node.field_file
     - node.type.report
   module:
-    - reliefweb_docstore
+    - reliefweb_files
 id: node.report.field_file
 field_name: field_file
 entity_type: node
