+++ config/reliefweb_files.settings.yml
@@ -1,0 +1,9 @@
+docstore_url: ''
+docstore_provider_uuid: ''
+docstore_api_key: ''
+docstore_prefix: reliefweb_
+local: true
+file_directory: attachments
+preview_directory: previews
+_core:
+  default_config_hash: p51VTBEaJE9L5iY3Imqr82p6ym3rNeOGMoKQ3VsimvA
