--- html/modules/custom/reliefweb_docstore/reliefweb_docstore.info.yml
@@ -1,5 +1,0 @@
-type: module
-name: ReliefWeb Docstore
-description: 'Provides integration with the OCHA Document Store.'
-package: reliefweb
-core_version_requirement: ^9
