--- html/modules/custom/reliefweb_docstore/reliefweb_docstore.routing.yml
@@ -1,14 +1,0 @@
-reliefweb_docstore.attachment.public:
-  path: '/attachments/{uuid}/{filename}'
-  defaults:
-    _controller: '\Drupal\reliefweb_docstore\Controller\FileDownloadController::downloadPublicAttachment'
-    _title: 'Download a public attachment'
-  requirements:
-    _permission: 'access content'
-reliefweb_docstore.attachment.private:
-  path: '/private/attachments/{uuid}/{filename}'
-  defaults:
-    _controller: '\Drupal\reliefweb_docstore\Controller\FileDownloadController::downloadPrivateAttachment'
-    _title: 'Download a private attachment'
-  requirements:
-    _permission: 'access content'
