+++ html/modules/custom/reliefweb_files/reliefweb_files.info.yml
@@ -1,0 +1,5 @@
+type: module
+name: ReliefWeb Files
+description: 'Handles file attachments and integration with the OCHA Document Store.'
+package: reliefweb
+core_version_requirement: ^9
