+++ html/modules/custom/reliefweb_files/reliefweb_files.routing.yml
@@ -1,0 +1,14 @@
+reliefweb_files.attachment.public:
+  path: '/attachments/{uuid}/{filename}'
+  defaults:
+    _controller: '\Drupal\reliefweb_files\Controller\FileDownloadController::downloadPublicAttachment'
+    _title: 'Download a public attachment'
+  requirements:
+    _permission: 'access content'
+reliefweb_files.attachment.private:
+  path: '/private/attachments/{uuid}/{filename}'
+  defaults:
+    _controller: '\Drupal\reliefweb_files\Controller\FileDownloadController::downloadPrivateAttachment'
+    _title: 'Download a private attachment'
+  requirements:
+    _permission: 'access content'
