+++ html/themes/custom/common_design_subtheme/components/rw-form/rw-form--widget.css
@@ -140,7 +140,7 @@ form .ief-form .fieldset-wrapper {
   background: var(--cd-reliefweb-brand-red--dark);
 }
 
-/* RW docstore */
+/* RW files */
 .rw-file-widget .tabledrag-toggle-weight-wrapper {
   display: none;
 }
