+++ html/themes/custom/common_design_subtheme/templates/rw-modules/reliefweb_docstore/reliefweb-file-list--interactive.html.twig
@@ -1,4 +1,4 @@
 {% set attributes = attributes.addClass('rw-attachment', 'rw-attachment--interactive') %}
 {% set title_attributes = title_attributes.addClass('rw-attachment__title') %}
 
-{% include '@reliefweb_docstore/reliefweb-file-list--interactive.html.twig' %}
+{% include '@reliefweb_files/reliefweb-file-list--interactive.html.twig' %}
