+++ html/themes/custom/common_design_subtheme/templates/rw-modules/reliefweb_docstore/reliefweb-file-list.html.twig
@@ -1,3 +1,3 @@
 {% set attributes = attributes.addClass('rw-attachment') %}
 
-{% include '@reliefweb_docstore/reliefweb-file-list.html.twig' %}
+{% include '@reliefweb_files/reliefweb-file-list.html.twig' %}
